requirements.txt
python-telegram-bot[socks]
telegraph
mitmproxy # should not installed if on jailbroken iOS
pytz
flask
python-dotenv
requests
paramiko